+++ src/scenes/browser-support-fixture.ts
@@ -38,85 +38,24 @@
 // for and would couple the gate to file-serving behavior that is
 // already covered by Vitest unit tests.
 //
+// The ctx validation and DOM mount/find/remove scaffolding is shared
+// with the loop verification fixture via `./fixture-support.ts`.
+//
 // The scene is `standalone: true` — it does not assume surrounding
 // composition context. `trailerSafe: false` — it has nothing
 // trailer-worthy to show.
 
-import { NAVIGATION_MODES } from '../runtime/navigation';
 import type { SceneModule } from '../runtime/scene';
-import type { WorkbenchSceneCtx } from '../runtime/scene-loader';
+import {
+  findFixtureElement,
+  isFixtureCtx,
+  mountFixtureElement,
+  removeFixtureElement,
+} from './fixture-support';
 
 const FIXTURE_TARGET_ATTR = 'data-pulsar-fixture-target';
 const FIXTURE_STATE_ATTR = 'data-pulsar-fixture-state';
 const FIXTURE_DURATION_SECONDS = 0.1;
-
-interface FixtureDomFactory {
-  createElement(tag: string): { setAttribute(name: string, value: string): void };
-}
-
-interface FixtureStageElement {
-  setAttribute(name: string, value: string): void;
-  removeAttribute(name: string): void;
-  appendChild?(node: unknown): unknown;
-  querySelector?(selector: string): {
-    setAttribute(name: string, value: string): void;
-    remove?(): void;
-  } | null;
-  // Real DOM elements expose `ownerDocument`; the fixture allocates
-  // DOM through the injected stage's owner document rather than the
-  // ambient global `document` so the scene stays pure against
-  // injected dependencies (ADR-008 #2). Optional so off-DOM test
-  // harnesses can supply a bare stage without a document.
-  readonly ownerDocument?: FixtureDomFactory | null;
-}
-
-// PUL-F012 / PUL-F022 / ADR-003: the runtime fills `ctx.stage`,
-// `ctx.mode`, and `ctx.gsap`. The fixture only reads those three
-// fields; the predicate narrows to that subset so a malformed ctx
-// (no `stage`, no `mode`, unknown `mode`, non-element `stage`) is
-// a no-op rather than a crash. Same defensive pattern the
-// placeholder scene uses, scoped to what this scene actually reads.
-// `stage` narrows to `FixtureStageElement | null` (not the
-// runtime's `StageElement | null`) so the lifecycle hooks can
-// read the optional `appendChild` / `querySelector` /
-// `ownerDocument` members without a per-hook type assertion.
-interface FixtureCtx {
-  readonly stage: FixtureStageElement | null;
-  readonly mode: WorkbenchSceneCtx['mode'];
-  readonly gsap: WorkbenchSceneCtx['gsap'];
-}
-
-const isStageShape = (stage: unknown): stage is FixtureStageElement | null => {
-  if (stage === null) return true;
-  if (typeof stage !== 'object') return false;
-  const candidate = stage as Partial<Record<'setAttribute' | 'removeAttribute', unknown>>;
-  return (
-    typeof candidate.setAttribute === 'function' && typeof candidate.removeAttribute === 'function'
-  );
-};
-
-const isGsapShape = (gsap: unknown): gsap is WorkbenchSceneCtx['gsap'] => {
-  if (gsap === null || typeof gsap !== 'object') return false;
-  return typeof (gsap as { timeline?: unknown }).timeline === 'function';
-};
-
-const isFixtureCtx = (value: unknown): value is FixtureCtx => {
-  if (typeof value !== 'object' || value === null) return false;
-  if (!('stage' in value) || !('mode' in value) || !('gsap' in value)) return false;
-  const { stage, mode, gsap } = value;
-  if (!isStageShape(stage)) return false;
-  if (typeof mode !== 'string' || !(NAVIGATION_MODES as readonly string[]).includes(mode)) {
-    return false;
-  }
-  return isGsapShape(gsap);
-};
-
-const findFixtureElement = (
-  stage: FixtureStageElement,
-): { setAttribute(name: string, value: string): void; remove?: () => void } | null => {
-  if (typeof stage.querySelector !== 'function') return null;
-  return stage.querySelector(`[${FIXTURE_TARGET_ATTR}]`);
-};
 
 export const browserSupportFixtureScene: SceneModule = {
   id: 'browser-support-fixture',
@@ -130,30 +69,13 @@ export const browserSupportFixtureScene: SceneModule = {
   standalone: true,
   trailerSafe: false,
   create: (ctx: unknown) => {
-    if (!isFixtureCtx(ctx)) return;
-    const stage = ctx.stage;
-    if (stage === null) return;
-    if (typeof stage.appendChild !== 'function') return;
-    // Allocate the fixture element through the stage's owner
-    // document — same dependency seam the runtime hands scenes via
-    // `ctx.stage`. Reaching for the ambient global `document` would
-    // bypass the explicit-dependency boundary the rest of the
-    // runtime enforces (ADR-008 #2; codex pre-push review, cycle
-    // 2). Off-DOM test harnesses that supply a stage without
-    // `ownerDocument` are a no-op rather than a crash.
-    const ownerDoc = stage.ownerDocument;
-    if (ownerDoc === undefined || ownerDoc === null) return;
-    if (typeof ownerDoc.createElement !== 'function') return;
-    const el = ownerDoc.createElement('div');
-    el.setAttribute(FIXTURE_TARGET_ATTR, '');
-    el.setAttribute(FIXTURE_STATE_ATTR, 'mounted');
-    stage.appendChild(el);
+    mountFixtureElement(ctx, FIXTURE_TARGET_ATTR, [[FIXTURE_STATE_ATTR, 'mounted']]);
   },
   timeline: (ctx: unknown) => {
     if (!isFixtureCtx(ctx)) return null;
     const stage = ctx.stage;
     if (stage === null) return null;
-    const el = findFixtureElement(stage);
+    const el = findFixtureElement(stage, FIXTURE_TARGET_ATTR);
     if (el === null) return null;
     // Build a real GSAP timeline through `ctx.gsap`. A `.call(...)` at
     // the timeline's end advances the fixture state, which the
@@ -170,12 +92,6 @@ export const browserSupportFixtureScene: SceneModule = {
     return tl;
   },
   cleanup: (ctx: unknown) => {
-    if (!isFixtureCtx(ctx)) return;
-    const stage = ctx.stage;
-    if (stage === null) return;
-    const el = findFixtureElement(stage);
-    if (el !== null && typeof el.remove === 'function') {
-      el.remove();
-    }
+    removeFixtureElement(ctx, FIXTURE_TARGET_ATTR);
   },
 };
